requirements.txt
streamlit
pandas
openpyxl
python-dotenv
openai==0.28
langchain
faiss-cpu